+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Correct calculation in `getClientRect` method of `Konva.Line` and `Konva.Container`.
 - Correct `toObject()` behaviour for node with attrs with extended native prototypes
+- Fixed bug for caching where buffer canvas is required
 
 ### Changed
 - Dragging now works much better. If your pointer is out of stage content dragging will still continue.